Print a more helpful error message if the input is not valid JSON
It appears to be a common issue that the `webpack --json` output can
include non-JSON content with certain plugins or loaders.

Trying to extract the JSON output automatically from the mixed input is
likely to be tricky, but we can at least provide a clearer diagnosis of
the problem.

Fixes #13
Fixes #19
Fixes #21
Fixes #30